+++ src/main/java/org/semux/config/Constants.java
@@ -34,7 +34,7 @@ public class Constants {
     /**
      * Version of this client.
      */
-    public static final String CLIENT_VERSION = "2.1.1";
+    public static final String CLIENT_VERSION = "2.2.0";
 
     /**
      * Algorithm name for the 256-bit hash.